+++ src/main/java/at/hannibal2/skyhanni/features/bazaar/BazaarData.kt
@@ -1,3 +1,11 @@
 package at.hannibal2.skyhanni.features.bazaar
 
-data class BazaarData(val apiName: String, val itemName: String, val sellPrice: Double, val buyPrice: Double, val buyMovingWeek: Int, val sellMovingWeek: Int)
+data class BazaarData(
+    val apiName: String,
+    val itemName: String,
+    val sellPrice: Double,
+    val buyPrice: Double,
+    val npcPrice: Double,
+    val buyMovingWeek: Int,
+    val sellMovingWeek: Int,
+)