+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 + Added **Colored Name** - Show the visitor name in the color of the rarity.
 + Added **Visitor Item Preview** - Show the base type for the required items next to new visitors (Note that some visitors may require any crop)
 + Added **Teleport Pad Compact Name** - Hide the 'Warp to' and 'No Destination' texts over teleport pads.
++ Added Money per Hour Advanced stats - Show not only Sell Offer price but also Instant Sell price and NPC Sell price (Suggestion: Enable Compact Price as well for this)
 
 
 ### Features from other Mods